fix bank deposit dup

+++ src/game.cpp
@@ -234,15 +234,24 @@ void game_select(const dpp::select_click_t &event) {
 	bot.log(dpp::ll_debug, std::to_string(event.command.usr.id) + ": " + custom_id);
 	if (custom_id == "withdraw" && p.in_bank && !event.values.empty()) {
 		std::vector<std::string> parts = dpp::utility::tokenize(event.values[0], ";");
-		db::query("DELETE FROM game_bank WHERE owner_id = ? AND item_desc = ? AND item_flags = ?", {event.command.usr.id, parts[0], parts[1]});
-		p.possessions.push_back(item{ .name = parts[0], .flags = parts[1] });
-		p.inv_change = true;
+		db::transaction();
+		auto rs = db::query("SELECT * FROM game_bank WHERE owner_id = ? AND item_desc = ? AND item_flags = ?", {event.command.usr.id, parts[0], parts[1]});
+		if (!rs.empty()) {
+			db::query("DELETE FROM game_bank WHERE owner_id = ? AND item_desc = ? AND item_flags = ?", {event.command.usr.id, parts[0], parts[1]});
+			p.possessions.push_back(item{.name = parts[0], .flags = parts[1]});
+			p.inv_change = true;
+		}
+		db::commit();
 		claimed = true;
 	} else if (custom_id == "deposit" && p.in_bank && !event.values.empty()) {
 		std::vector<std::string> parts = dpp::utility::tokenize(event.values[0], ";");
-		db::query("INSERT INTO game_bank (owner_id, item_desc, item_flags ) VALUES(?,?,?)", {event.command.usr.id, parts[0], parts[1]});
-		p.drop_possession(item{ .name = parts[0], .flags = parts[1] });
-		p.inv_change = true;
+		db::transaction();
+		if (p.has_possession(parts[0])) {
+			db::query("INSERT INTO game_bank (owner_id, item_desc, item_flags ) VALUES(?,?,?)", {event.command.usr.id, parts[0], parts[1]});
+			p.drop_possession(item{.name = parts[0], .flags = parts[1]});
+			p.inv_change = true;
+		}
+		db::commit();
 		claimed = true;
 	} else if (custom_id == "fight_pvp" && p.in_pvp_picker && !event.values.empty()) {
 		dpp::snowflake user(event.values[0]);
